refactor: change return value for forces methods

+++ src/main/kotlin/viewmodel/placement/ForceAtlas2VertexLayout.kt
@@ -4,10 +4,10 @@ import viewmodel.graphs.VertexViewModel
 
 data class ForceAtlas2VertexLayout<T>(
     val vertex: VertexViewModel<T>,
-    var dltX: Double = 0.0,
-    var dltY: Double = 0.0,
+    var dltX: Float = 0f,
+    var dltY: Float = 0f,
 ) {
-    fun addForces(xF: Double, yF: Double) {
+    fun addForces(xF: Float, yF: Float) {
         dltX += xF
         dltY += yF
     }
